Аdd solution
