+++ Additional/Homeworks/hmw2/Kindle/Source.cpp
@@ -1,0 +1,6 @@
+#include"Kindle.h"
+
+int main() {
+	Kindle env;
+	env.display();
+}
