--- Kindle/Source.cpp
@@ -1,6 +1,0 @@
-#include"Kindle.h"
-
-int main() {
-	Kindle env;
-	env.display();
-}
